Refactor current `kind` e2e test location
